Fetch maximum duration for the chosen role
The value for `MaxSessionDuration` depends on the role chosen by the
user. We can therefore not assume 12 hours to be the maximum.
In order to acquire the information necessary, we need to assume the
chosen role temporarily (using the minimum session length of 15
minutes) and contact IAM. We then ask the user for the desired session
length and present the user with the previously acquired maximum.

Fixes #93