Update README.md
Add a note to the docker installation section surrounding
the usage of docker behind a corporate proxy. 

The vanilla docker run command did not work for me
without the setting of the proxy environment variables.
It took me quite some time to figure out why the
container was simply 'hanging'. This addition to the
README.md might help others, who are
using the docker aws-azure-login route, troubleshoot
faster.