Allow store of Chromium data per Azure AD tenant
This change allows for storing of the Chromium user data on a per Azure AD tenant basis for opted-in AWS config named profiles. The point here is that some users interact with more than one Azure AD tenant and want to leverage the remember me functionality in an unsurprising way. With the Chromium user data stored in one fixed location, authentication attempts that cross back and forth between Azure AD tenants aren't guaranteed to work. This feature ensures a strict separation between the stored user data on a per tenant basis and ensures remembered login items are never blindly used in an incorrect fashion.